requirements.txt
aiosignal==1.2.0
algoliasearch==2.6.0
algoliasearch-django==2.0.0
asgiref==3.4.1
attrs==21.2.0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-common-helpers==0.9.2
django-crontab==0.7.1
django-environ==0.8.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.0
drf-yasg==1.20.0
environ==1.0
frozenlist==1.2.0
google-api-core==1.31.4
google-auth==1.35.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.1.0
google-cloud-logging==2.7.0
google-cloud-secret-manager==2.0.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.1
grpcio-status==1.41.1
gunicorn==20.1.0
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonpickle==0.9.6
jsonschema==4.2.1
libcst==0.3.23
MarkupSafe==2.0.1
msgpack==1.0.3
multidict==5.2.0
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==21.2
postgis==1.0.4
proto-plus==1.19.7
protobuf==3.19.1
psycopg2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1
typing-inspect==0.7.1
uritemplate==4.1.1
urllib3==1.26.7
yarl==1.7.2
zipp==3.6.0
